Clarify what's redacted
While I appreciated the ;) nature of the current sample text, it reads in a slightly confusing way, as if it is text declaring the following text to be redacted, not that the text within the curly brackets _is_ redacted. For users not familiar with markup of this type, this text might be more straightforward.